requirements.txt
asgiref==3.6.0
astroid==2.15.6
async-timeout==4.0.2
billiard==4.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
Django==4.1.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
kombu==5.3.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
pyarrow==12.0.1
pylance==0.5.8
pylint==2.17.5
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0.1
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
wrapt==1.15.0
